working chat v4

+++ backend/app/models/score.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, ForeignKey
+from sqlalchemy import Column, Integer, ForeignKey, Float
 from sqlalchemy.orm import relationship
 from app.core.db import Base
 
@@ -9,10 +9,10 @@ class Score(Base):
 
     profile_id = Column(Integer, ForeignKey("profiles.id"), unique=True)
 
-    math_score = Column(Integer, default=0)
-    tech_score = Column(Integer, default=0)
-    arts_score = Column(Integer, default=0)
-    commerce_score = Column(Integer, default=0)
-    science_score = Column(Integer, default=0)
+    math_score = Column(Float, default=0)
+    tech_score = Column(Float, default=0)
+    arts_score = Column(Float, default=0)
+    commerce_score = Column(Float, default=0)
+    science_score = Column(Float, default=0)
     
     profile = relationship("UserProfile", back_populates="score")
